my_project/requirements.txt
dbt-core>=1.10.15
dbt-postgres>=1.9.1
dotenv>=0.9.9
pandas>=2.3.3
polars>=1.35.1
psycopg2-binary>=2.9.11
pyodbc>=5.3.0
pypyodbc>=1.3.6
pytz>=2025.2
rich>=14.2.0